Skip DELETE of NVT prefs if table already truncated

+++ src/manage_sql_configs.c
@@ -1694,21 +1694,23 @@ check_config_families ()
 /**
  * @brief Add/replace an NVT preference.
  *
- * @param[in]  name    The name of the preference.
- * @param[in]  value   The value of the preference.
+ * @param[in]  name       The name of the preference.
+ * @param[in]  value      The value of the preference.
+ * @param[in]  truncated  Whether nvt_preferences was truncated beforehand.
  */
 void
-manage_nvt_preference_add (const char* name, const char* value)
+manage_nvt_preference_add (const char* name, const char* value, int truncated)
 {
   gchar* quoted_name = sql_quote (name);
   gchar* quoted_value = sql_quote (value);
 
   if (strcmp (name, "port_range"))
     {
-      if (sql_int ("SELECT EXISTS"
-                   "  (SELECT * FROM nvt_preferences"
-                   "   WHERE name = '%s')",
-                   quoted_name))
+      if ((truncated == 0)
+          && sql_int ("SELECT EXISTS"
+                      "  (SELECT * FROM nvt_preferences"
+                      "   WHERE name = '%s')",
+                      quoted_name))
         sql ("DELETE FROM nvt_preferences WHERE name = '%s';", quoted_name);
 
       sql ("INSERT into nvt_preferences (name, value)"
